0.3.0 Histogram

+++ libraries/Histogram/examples/hist_test_val/hist_test_val.ino
@@ -16,6 +16,7 @@ float b[] = {
 
 Histogram hist(16, b);
 
+
 void setup()
 {
   Serial.begin(115200);
@@ -23,6 +24,7 @@ void setup()
   Serial.print("Histogram version: ");
   Serial.println(HISTOGRAM_LIB_VERSION);
 }
+
 
 void loop()
 {
@@ -39,8 +41,8 @@ void loop()
   Serial.print("\t");
   
   float sum = 0;
-  int i = 0;
-  for (i = 0; i< hist.size(); i++)
+  uint16_t i = 0;
+  for (i = 0; i < hist.size(); i++)
   {
     sum += hist.frequency(i);
     Serial.print(sum, 2);
@@ -55,4 +57,5 @@ void loop()
   delay(10);
 }
 
-// END OF FILE
+
+// -- END OF FILE --
